Move data out of elaphos and into RDS files in data directory.
Formerly, elaphos was primarily a data package, and the data for these
experiments was housed there. However, there are data in that package
that are not for public release, making it impossible to share the
elaphos package alongside the publication.

Thus, we've moved all the data into stand-alone data objects in the
data directory.